refactor: move AI financial route logic into dedicated controller methods #541
